+++ CHANGELOG.md
@@ -3,6 +3,22 @@ aegir3 CHANGELOG
 
 This file is used to list changes made in each version of the aegir3 cookbook.
 
-0.1.0 (2014-06-24)
+0.1.0 (2015-06-07)
 ------------------
 - Dieter Blomme - First version, adaptation of the aegir2 cookbook
+
+0.1.1 (2015-06-08)
+------------------
+- Dieter Blomme - Code style conformation
+- Dieter Blomme - Nginx install works
+
+0.1.2 (2015-06-15)
+------------------
+- Dieter Blomme - Update MySQL for new mysql cookbook
+- Dieter Blomme - Separation MySQL install from aegir install
+- Dieter Blomme - Add db_host parameter for remote install
+- Dieter Blomme - Enable tasks queue by default
+- Dieter Blomme - Add default for mysql root password
+- Dieter Blomme - Only use mysql_secure for localhost
+- Dieter Blomme - Use Nginx cookbook
+- Dieter Blomme - Various rubocop en foodcritic fixes